refactor: update chain handling and remove unused code across components
