+++ app/components/common/format/Money.vue
@@ -1,6 +1,4 @@
 <script lang="ts" setup>
-import type { Prefix } from '@kodadot1/static'
-import { chainPropListOf } from '@/utils/chain'
 import { formatAmountWithRound } from '@/utils/format/balance'
 
 const props = withDefaults(
@@ -9,28 +7,19 @@ const props = withDefaults(
     inline: boolean
     hideUnit?: boolean
     unitSymbol?: string
-    prefix?: Prefix
     round?: number
   }>(),
   {
     value: 0,
     unitSymbol: '',
-    prefix: undefined,
   },
 )
 
 const { decimals, chainSymbol } = useChain()
 
-const tokenDecimals = computed(() =>
-  props.prefix ? chainPropListOf(props.prefix).tokenDecimals : decimals.value,
-)
-const displayUnit = computed(
-  () =>
-    props.unitSymbol
-    || (props.prefix ? chainPropListOf(props.prefix).tokenSymbol : chainSymbol.value),
-)
+const displayUnit = computed(() => props.unitSymbol || chainSymbol.value)
 const finalValue = computed(() =>
-  formatAmountWithRound(props.value, tokenDecimals.value, props.round),
+  formatAmountWithRound(props.value, decimals.value, props.round),
 )
 </script>
