+++ app/components/drop/DropCollectedBy.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
-import type { Prefix } from '@kodadot1/static'
+import type { AssetHubChain } from '~/plugins/sdk.client'
 import { fetchOdaCollectionOwners } from '@/services/oda'
 
 const props = withDefaults(
   defineProps<{
-    chain: Prefix
+    chain: AssetHubChain
     collectionId: string
     maxAddressCount?: number
     size?: 'small' | 'medium'
