+++ app/utils/network.ts
@@ -1,5 +1,5 @@
 import type { Prefix } from '@kodadot1/static'
-import { vmOf } from './chain'
+import { vmOf } from '@/composables/useVm'
 
 export const isEvm = (prefix: Prefix) => vmOf(prefix) === 'EVM'
 export const isSub = (prefix: Prefix) => vmOf(prefix) === 'SUB'
